# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: LB CodeBase API
  version: 2.0.0
  description: MariaDB-backed public content and authenticated administration API.
servers:
  - url: /api/v2
...
paths:
...
  /projects:
    get:
      tags: [Public]
      operationId: listProjects
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }, { name: category, in: query, schema: { type: string } }]
      responses:
        '200': { description: Published projects, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectPage' } } } }
...
  /team:
    get:
      tags: [Public]
      operationId: listTeamMembers
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }]
      responses:
        '200': { description: Active team members }
...
  /blog:
    get:
      tags: [Public]
      operationId: listBlogPosts
      parameters: [{ $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }]
      responses:
        '200': { description: Published blog posts }
...
  /messages:
    post:
      tags: [Public]
      operationId: createContactMessage
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ContactMessageInput' } } }
      responses:
        '201': { description: Message accepted }
        '422': { $ref: '#/components/responses/Validation' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /admin/projects:
    get:
      tags: [Admin]
      operationId: adminListProjects
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: All projects }, '401': { $ref: '#/components/responses/Unauthorized' } }
    post:
      tags: [Admin]
      operationId: adminCreateProject
      security: [{ sessionCookie: [], csrfToken: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectInput' } } } }
      responses: { '201': { description: Project created }, '403': { $ref: '#/components/responses/Forbidden' }, '422': { $ref: '#/components/responses/Validation' } }
...
  /admin/team:
    get:
      tags: [Admin]
      operationId: adminListTeam
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: All team profiles } }
    post:
      tags: [Admin]
      operationId: adminCreateTeamMember
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses: { '201': { description: Team profile created }, '422': { $ref: '#/components/responses/Validation' } }
...
  /admin/blog:
    get:
      tags: [Admin]
      operationId: adminListBlogPosts
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: All blog posts } }
    post:
      tags: [Admin]
      operationId: adminCreateBlogPost
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses: { '201': { description: Blog post created }, '422': { $ref: '#/components/responses/Validation' } }
...
  /admin/services:
    get:
      tags: [Admin]
      operationId: adminListServices
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: All services } }
    post:
      tags: [Admin]
      operationId: adminCreateService
      security: [{ sessionCookie: [], csrfToken: [] }]
      responses: { '201': { description: Service created } }
...
  /admin/messages:
    get:
      tags: [Admin]
      operationId: adminListMessages
      security: [{ sessionCookie: [] }]
      responses: { '200': { description: Paginated contact workflow } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: lb_session }
    csrfToken: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 200 } }
...
  schemas:
...
    Project:
      type: object
      required: [id, title, slug, status, technologies, contributors, version]
      properties: { id: { type: string }, title: { type: string }, slug: { type: string }, thumbnail: { type: string }, thumbnailMediaId: { type: [string, 'null'] }, gallery: { type: array, items: { type: string } }, galleryMediaIds: { type: array, items: { type: string } }, status: { enum: [draft, published, archived] }, technologies: { type: array, items: { type: string } }, contributors: { type: array, items: { type: object } }, version: { type: integer } }
    ProjectInput:
      type: object
      additionalProperties: false
      required: [title]
      properties: { title: { type: string, maxLength: 240 }, slug: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }, thumbnail: { type: string }, thumbnailMediaId: { type: [string, 'null'] }, gallery: { type: array, maxItems: 30, items: { type: string } }, galleryMediaIds: { type: array, maxItems: 30, items: { type: string } }, status: { enum: [draft, published, archived], default: draft }, technologies: { type: array, maxItems: 100, items: { type: string } }, contributors: { type: array, maxItems: 50, items: { type: object, required: [memberId, role], properties: { memberId: { type: string }, role: { type: string } } } }, version: { type: integer, minimum: 1 } }
    ProjectPage:
      type: object
      required: [items, page, limit, total, pageCount]
      properties: { items: { type: array, items: { $ref: '#/components/schemas/Project' } }, page: { type: integer }, limit: { type: integer }, total: { type: integer }, pageCount: { type: integer } }
...
    ContactMessageInput:
      type: object
      additionalProperties: false
      required: [name, email, subject, message]
      properties: { name: { type: string, minLength: 2, maxLength: 120 }, email: { type: string, format: email, maxLength: 254 }, subject: { type: string, maxLength: 160 }, message: { type: string, minLength: 10, maxLength: 5000 }, memberId: { type: string }, website: { type: string, maxLength: 0 } }
    Error:
      type: object
      required: [success, error]
      properties:
        success: { const: false }
        error:
          type: object
          required: [code, message, requestId]
          properties: { code: { type: string }, message: { type: string }, requestId: { type: string }, fields: { type: object, additionalProperties: { type: array, items: { type: string } } } }
  responses:
    Unauthorized: { description: Authentication required or session expired, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    Forbidden: { description: CSRF or RBAC rejected the request, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Validation: { description: Input validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    RateLimited: { description: Rate limit exceeded, headers: { Retry-After: { schema: { type: integer } } }, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...